Switch to package-lock.json and refactor tests
