+++ package.json
@@ -56,22 +56,22 @@
   ],
   "scripts": {
     "now-build": "build-storybook -c .storybook -o public",
-    "build": "tsdx build --format=umd,esm,cjs&&yarn sass scss/styles.scss dist/css/styles.css",
+    "build": "tsdx build --format=umd,esm,cjs&&npm run sass scss/styles.scss dist/css/styles.css",
     "dev": "tsdx watch --format=esm",
-    "start": "yarn dev",
+    "start": "npm run dev",
     "test": "cross-env CI=1 TZ=Europe/Stockholm vitest",
     "test:ci": "cross-env CI=1 TZ=Europe/Stockholm vitest",
-    "test:coverage": "yarn test --coverage",
-    "test:all": "echo 'Running tests...' && yarn bundlesize && yarn test:coverage && echo 'All tests passed! 🤘'",
+    "test:coverage": "npm run test --coverage",
+    "test:all": "echo 'Running tests...' && npm run bundlesize && npm run test:coverage && echo 'All tests passed! 🤘'",
     "test:watch": "cross-env TZ=Europe/Stockholm vitest --watch",
-    "deduplicate": "npx yarn-deduplicate -s fewer yarn.lock",
-    "lint": "yarn eslint src --ext .js,.ts,.jsx,.tsx",
-    "lint:fix": "yarn lint --fix",
+    "deduplicate": "npx npm run-deduplicate -s fewer npm run.lock",
+    "lint": "npm run eslint src --ext .js,.ts,.jsx,.tsx",
+    "lint:fix": "npm run lint --fix",
     "prettier": "prettier --write .",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook -c .storybook -o .out",
     "chromatic": "chromatic",
-    "prepack": "yarn build",
+    "prepack": "npm run build",
     "prepare": "husky install",
     "pre-commit": "lint-staged"
   },
@@ -109,6 +109,7 @@
     "@types/react-dom": "^17.0.11",
     "@typescript-eslint/eslint-plugin": "^5.12.1",
     "@typescript-eslint/parser": "^5.12.1",
+    "@vitest/ui": "^0.5.7",
     "awesome-typescript-loader": "^5.2.1",
     "babel-jest": "^27.5.1",
     "babel-loader": "^8.2.3",
@@ -126,6 +127,7 @@
     "eslint-plugin-storybook": "^0.5.7",
     "eslint-plugin-testing-library": "^5.0.5",
     "husky": "^7.0.0",
+    "jest-environment-jsdom-sixteen": "^2.0.0",
     "lint-staged": "^12.3.4",
     "prettier": "^2.5.1",
     "react": "^17.0.2",
@@ -140,7 +142,7 @@
     "tsdx": "0.14.1",
     "typescript": "^4.5.5",
     "vite": "^2.8.4",
-    "vitest": "0.5.4",
+    "vitest": "^0.5.7",
     "webpack": "5.69.1"
   },
   "dependencies": {
