+++ src/DateTime.test.tsx
@@ -1789,1459 +1789,4 @@ describe("DateTime", () => {
       expect(await screen.findByTestId("year-picker")).toBeVisible();
     });
   });
-
-  describe("end-to-end", () => {
-    it("should switch through to year mode, and pick a specific date through various pickers", async () => {
-      mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-      // Arrange
-      render(
-        <>
-          <label htmlFor="some-id">Some Field</label>
-          <DateTime
-            id="some-id"
-            dateFormat={FULL_DATE_FORMAT}
-            timeFormat={false}
-          />
-        </>
-      );
-
-      const element = await screen.findByLabelText("Some Field");
-
-      // Act
-      userEvent.click(element);
-      expect(element).toHaveFocus();
-
-      {
-        const picker = await screen.findByTestId("day-picker");
-        expect(picker).toBeVisible();
-      }
-
-      // day -> month picker
-      {
-        const switcher = await screen.findByTestId("day-mode-switcher");
-        expect(switcher).toBeVisible();
-
-        userEvent.click(switcher);
-
-        // Assert
-        const picker = await screen.findByTestId("month-picker");
-        expect(picker).toBeVisible();
-      }
-
-      // month -> year picker
-      {
-        const switcher = await screen.findByTestId("month-mode-switcher");
-        expect(switcher).toBeVisible();
-
-        userEvent.click(switcher);
-
-        // Assert
-        const picker = await screen.findByTestId("year-picker");
-        expect(picker).toBeVisible();
-      }
-
-      {
-        // click a year (switch to month picker)
-        userEvent.click(await screen.findByText("2020"));
-
-        const picker = await screen.findByTestId("month-picker");
-        expect(picker).toBeVisible();
-      }
-
-      {
-        // click a month (switch to day picker)
-        userEvent.click(await screen.findByText("Feb"));
-
-        const picker = await screen.findByTestId("day-picker");
-        expect(picker).toBeVisible();
-      }
-
-      // click a day
-      userEvent.click(await screen.findByText("11"));
-
-      // Assert
-      expect(await screen.findByLabelText("Some Field")).toHaveValue(
-        "02/11/2020"
-      );
-    });
-
-    it("should switch through to year mode, and pick a specific date/time through various pickers", async () => {
-      mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-      // Arrange
-      render(
-        <>
-          <label htmlFor="some-id">Some Field</label>
-          <DateTime
-            id="some-id"
-            dateFormat={FULL_DATE_FORMAT}
-            timeFormat={`${FORMATS.HOUR}:${FORMATS.MINUTE}:${FORMATS.SECOND}.${FORMATS.MILLISECOND} ${FORMATS.AM_PM}`}
-          />
-        </>
-      );
-
-      const element = await screen.findByLabelText("Some Field");
-
-      // Act
-      userEvent.click(element);
-      expect(element).toHaveFocus();
-
-      {
-        const picker = await screen.findByTestId("day-picker");
-        expect(picker).toBeVisible();
-      }
-
-      // day -> month picker
-      {
-        const switcher = await screen.findByTestId("day-mode-switcher");
-        expect(switcher).toBeVisible();
-
-        userEvent.click(switcher);
-
-        // Assert
-        const picker = await screen.findByTestId("month-picker");
-        expect(picker).toBeVisible();
-      }
-
-      // month -> year picker
-      {
-        const switcher = await screen.findByTestId("month-mode-switcher");
-        expect(switcher).toBeVisible();
-
-        userEvent.click(switcher);
-
-        // Assert
-        const picker = await screen.findByTestId("year-picker");
-        expect(picker).toBeVisible();
-      }
-
-      {
-        // click a year (switch to month picker)
-        userEvent.click(await screen.findByText("2020"));
-
-        const picker = await screen.findByTestId("month-picker");
-        expect(picker).toBeVisible();
-      }
-
-      {
-        // click a month (switch to day picker)
-        userEvent.click(await screen.findByText("Feb"));
-
-        const picker = await screen.findByTestId("day-picker");
-        expect(picker).toBeVisible();
-      }
-
-      // Switch to time mode
-      {
-        const switcher = await screen.findByTestId("day-to-time-mode-switcher");
-        expect(switcher).toBeVisible();
-
-        userEvent.click(switcher);
-      }
-
-      {
-        // Assert
-        const picker = await screen.findByTestId("time-picker");
-        expect(picker).toBeVisible();
-
-        expect(picker.textContent?.replace(/[^\w/]+/g, "")).toMatch(
-          /02\/01\/2020120000000AM/i
-        );
-
-        // Click to change the time
-        const upArrows = await screen.findAllByText("▲");
-        expect(upArrows?.length).toBe(5);
-
-        const downArrows = await screen.findAllByText("▼");
-        expect(downArrows?.length).toBe(5);
-
-        // Increase hours from 12 to 1
-        userEvent.click(upArrows[0]);
-
-        // Increase minutes from 00 to 05
-        for (let i = 0; i < 5; i++) {
-          userEvent.click(upArrows[1]);
-        }
-
-        // Increase seconds from 00 to 35
-        for (let i = 0; i < 35; i++) {
-          userEvent.click(upArrows[2]);
-        }
-
-        // Increase milliseconds from 000 to 321
-        for (let i = 0; i < 321; i++) {
-          userEvent.click(upArrows[3]);
-        }
-
-        // Change from AM to PM
-        userEvent.click(upArrows[4]);
-
-        expect(picker.textContent?.replace(/[^\w/]+/g, "")).toMatch(
-          /02\/01\/202010535321PM/i
-        );
-
-        // Change from PM to AM
-        userEvent.click(upArrows[4]);
-
-        expect(picker.textContent?.replace(/[^\w/]+/g, "")).toMatch(
-          /02\/01\/202010535321AM/i
-        );
-
-        // Change from AM to PM
-        userEvent.click(upArrows[4]);
-
-        expect(picker.textContent?.replace(/[^\w/]+/g, "")).toMatch(
-          /02\/01\/202010535321PM/i
-        );
-      }
-
-      // Switch back to day mode
-      {
-        const switcher = await screen.findByTestId("time-mode-switcher");
-        expect(switcher).toBeVisible();
-
-        userEvent.click(switcher);
-      }
-
-      // click a day
-      userEvent.click(await screen.findByText("11"));
-
-      // Assert
-      expect(await screen.findByLabelText("Some Field")).toHaveValue(
-        "02/11/2020 1:05:35.321 PM"
-      );
-    });
-
-    describe("events", () => {
-      describe("onChange", () => {
-        it("should not trigger onChange Date when opening w/ Date", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onChange={handleChange}
-                value={new Date(2019, 0, 16, 12, 1, 12, 34)}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("01/16/2019");
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          expect(handleChange).toHaveBeenCalledTimes(0);
-        });
-
-        it("should not trigger onChange Date when clicking same date", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onChange={handleChange}
-                value={new Date(2019, 0, 16, 12, 1, 12, 34)}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("01/16/2019");
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-          const someDay = await screen.findByText("16");
-          expect(someDay).toBeVisible();
-
-          // Pick date
-          userEvent.click(someDay);
-
-          expect(element).toHaveValue("01/16/2019");
-
-          expect(handleChange).toHaveBeenCalledTimes(0);
-        });
-
-        it("should trigger onChange Date when picking a first date", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onChange={handleChange}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-          const someDay = await screen.findByText("16");
-          expect(someDay).toBeVisible();
-
-          // Pick date
-          userEvent.click(someDay);
-
-          expect(element).toHaveValue("01/16/2019");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith(
-            new Date(2019, 0, 16, 0, 0, 0, 0)
-          );
-        });
-
-        it("should trigger onChange Date when picking a new date", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onChange={handleChange}
-                value={new Date(2019, 0, 16, 12, 1, 12, 34)}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("01/16/2019");
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-          const someDay = await screen.findByText("17");
-          expect(someDay).toBeVisible();
-
-          // Pick date
-          userEvent.click(someDay);
-
-          expect(element).toHaveValue("01/17/2019");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith(
-            new Date(2019, 0, 17, 12, 1, 12, 34)
-          );
-        });
-
-        it("should trigger onChange input string when picking a date", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-          const someDay = await screen.findByText("16");
-          expect(someDay).toBeVisible();
-
-          // Pick date
-          userEvent.click(someDay);
-
-          expect(element).toHaveValue("01/16/2019");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("01/16/2019");
-        });
-
-        it("should trigger onChange input string when picking a date/time", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={FULL_TIME_FORMAT}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-              />
-            </>
-          );
-
-          expect(await screen.findByLabelText("Some Field")).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(await screen.findByLabelText("Some Field"));
-
-          // Assert
-          expect(await screen.findByTestId("day-picker")).toBeVisible();
-          expect(await screen.findByText("16")).toBeVisible();
-
-          // Pick date
-          userEvent.click(await screen.findByText("16"));
-
-          expect(await screen.findByLabelText("Some Field")).toHaveValue(
-            "01/16/2019 12:00 AM"
-          );
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("01/16/2019 12:00 AM");
-        });
-
-        it("should trigger onChange input string when increasing time by one step", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={false}
-                timeFormat={FULL_TIME_FORMAT}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-              />
-            </>
-          );
-
-          expect(await screen.findByLabelText("Some Field")).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(await screen.findByLabelText("Some Field"));
-
-          // Assert
-          expect(await screen.findByTestId("time-picker")).toBeVisible();
-
-          // Click to change the time
-          const upArrows = await screen.findAllByText("▲");
-          expect(upArrows?.length).toBe(3);
-          const downArrows = await screen.findAllByText("▼");
-          expect(downArrows?.length).toBe(3);
-
-          // Increase hours from 12 to 1
-          userEvent.click(upArrows[0]);
-
-          expect(await screen.findByLabelText("Some Field")).toHaveValue(
-            "1:00 AM"
-          );
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("1:00 AM");
-        });
-
-        it("should trigger onChange input string when increasing time by 3 hour increment", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={false}
-                timeFormat={FULL_TIME_FORMAT}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-                timeConstraints={{
-                  hours: {
-                    step: 3,
-                  },
-                }}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("time-picker")).toBeVisible();
-
-          // Click to change the time
-          const upArrows = await screen.findAllByText("▲");
-          expect(upArrows?.length).toBe(3);
-
-          const downArrows = await screen.findAllByText("▼");
-          expect(downArrows?.length).toBe(3);
-
-          // Increase hours from 12 to 3
-          userEvent.click(upArrows[0]);
-
-          expect(element).toHaveValue("3:00 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("3:00 AM");
-
-          // Decrease hours from 3 to 12
-          userEvent.click(downArrows[0]);
-
-          expect(element).toHaveValue("12:00 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(2);
-          expect(handleChange).toHaveBeenCalledWith("12:00 AM");
-        });
-
-        it("should trigger onChange input string when increasing time by 15 min increment", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={false}
-                timeFormat={FULL_TIME_FORMAT}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-                timeConstraints={{
-                  minutes: {
-                    step: 15,
-                  },
-                }}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("time-picker")).toBeVisible();
-
-          // Click to change the time
-          const upArrows = await screen.findAllByText("▲");
-          expect(upArrows?.length).toBe(3);
-
-          const downArrows = await screen.findAllByText("▼");
-          expect(downArrows?.length).toBe(3);
-
-          // Increase minutes from 0 to 15
-          userEvent.click(upArrows[1]);
-
-          expect(element).toHaveValue("12:15 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("12:15 AM");
-
-          // Decrease minutes from 15 to 0
-          userEvent.click(downArrows[1]);
-
-          expect(element).toHaveValue("12:00 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(2);
-          expect(handleChange).toHaveBeenCalledWith("12:00 AM");
-        });
-
-        it("should trigger onChange input string when increasing time by 30 second increment", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={false}
-                timeFormat={`${FORMATS.HOUR}:${FORMATS.MINUTE}:${FORMATS.SECOND} ${FORMATS.AM_PM}`}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-                timeConstraints={{
-                  seconds: {
-                    step: 30,
-                  },
-                }}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("time-picker")).toBeVisible();
-
-          // Click to change the time
-          const upArrows = await screen.findAllByText("▲");
-          expect(upArrows?.length).toBe(4);
-
-          const downArrows = await screen.findAllByText("▼");
-          expect(downArrows?.length).toBe(4);
-
-          // Increase seconds from 0 to 30
-          userEvent.click(upArrows[2]);
-
-          expect(element).toHaveValue("12:00:30 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("12:00:30 AM");
-
-          // Decrease seconds from 30 to 0
-          userEvent.click(downArrows[2]);
-
-          expect(element).toHaveValue("12:00:00 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(2);
-          expect(handleChange).toHaveBeenCalledWith("12:00:00 AM");
-        });
-
-        it("should trigger onChange input string when increasing time by 10 milliseconds increment", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={false}
-                timeFormat={`${FORMATS.HOUR}:${FORMATS.MINUTE}:${FORMATS.SECOND}.${FORMATS.MILLISECOND} ${FORMATS.AM_PM}`}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-                timeConstraints={{
-                  milliseconds: {
-                    step: 10,
-                  },
-                }}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("time-picker")).toBeVisible();
-
-          // Click to change the time
-          const upArrows = await screen.findAllByText("▲");
-          expect(upArrows?.length).toBe(5);
-
-          const downArrows = await screen.findAllByText("▼");
-          expect(downArrows?.length).toBe(5);
-
-          // Increase milliseconds from 0 to 10
-          userEvent.click(upArrows[3]);
-
-          expect(element).toHaveValue("12:00:00.010 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("12:00:00.010 AM");
-
-          // Decrease milliseconds from 10 to 0
-          userEvent.click(downArrows[3]);
-
-          expect(element).toHaveValue("12:00:00.000 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(2);
-          expect(handleChange).toHaveBeenCalledWith("12:00:00.000 AM");
-        });
-
-        it('should trigger onChange input string once holding down the "up" seconds for a bit of time', async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={false}
-                timeFormat={`${FORMATS.HOUR}:${FORMATS.MINUTE}:${FORMATS.SECOND} ${FORMATS.AM_PM}`}
-                onChange={handleChange}
-                dateTypeMode="input-format"
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("time-picker")).toBeVisible();
-
-          // Click to change the time
-          const upArrows = await screen.findAllByText("▲");
-          expect(upArrows?.length).toBe(4);
-
-          const downArrows = await screen.findAllByText("▼");
-          expect(downArrows?.length).toBe(4);
-
-          // Use fake timers to run for x seconds
-          vi.useFakeTimers();
-
-          fireEvent.mouseDown(upArrows[2]);
-
-          // Increase seconds
-          act(() => {
-            // Fast forward and exhaust only currently pending timers
-            // (but not any new timers that get created during that process)
-            vi.runOnlyPendingTimers();
-
-            // Fast forward and exhaust only currently pending timers
-            // (but not any new timers that get created during that process)
-            vi.runOnlyPendingTimers();
-
-            // Fast forward and exhaust only currently pending timers
-            // (but not any new timers that get created during that process)
-            vi.runOnlyPendingTimers();
-
-            // Fast forward and exhaust only currently pending timers
-            // (but not any new timers that get created during that process)
-            vi.runOnlyPendingTimers();
-          });
-
-          fireEvent.mouseUp(upArrows[2]);
-
-          // Re-enable real timers
-          vi.useRealTimers();
-
-          expect(element).toHaveValue("12:00:04 AM");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith("12:00:04 AM");
-        });
-
-        it("should trigger onChange utc-ms-timestamp when picking a date", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleChange = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onChange={handleChange}
-                dateTypeMode="utc-ms-timestamp"
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-          const someDay = await screen.findByText("16");
-          expect(someDay).toBeVisible();
-
-          // Pick date
-          userEvent.click(someDay);
-
-          expect(element).toHaveValue("01/16/2019");
-
-          expect(handleChange).toHaveBeenCalledTimes(1);
-          expect(handleChange).toHaveBeenCalledWith(
-            new Date(2019, 0, 16, 0, 0, 0, 0).getTime()
-          );
-        });
-      });
-
-      describe("onFocus", () => {
-        it("should trigger onFocus when tabbed in", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleFocus = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onFocus={handleFocus}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          expect(document.body).toHaveFocus();
-
-          // Open picker
-          userEvent.tab();
-          expect(element).toHaveFocus();
-
-          // Should have triggered "onFocus"
-          expect(handleFocus).toHaveBeenCalledTimes(1);
-        });
-
-        it("should trigger onFocus when clicked in", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleFocus = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onFocus={handleFocus}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          userEvent.click(element);
-
-          // Should have triggered "onFocus"
-          expect(handleFocus).toHaveBeenCalledTimes(1);
-        });
-      });
-
-      describe("onBlur", () => {
-        it("should trigger onBlur with no value when tabbed out with no value", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleBlur = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onBlur={handleBlur}
-              />
-              <label htmlFor="another-id">Another Field</label>
-              <input id="another-id" type="text" />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          expect(document.body).toHaveFocus();
-
-          // Tab in
-          userEvent.tab();
-          expect(element).toHaveFocus();
-
-          // Tab out
-          userEvent.tab();
-
-          // Should have triggered "onBlur"
-          expect(handleBlur).toHaveBeenCalledTimes(1);
-          expect(element).not.toHaveFocus();
-          expect(handleBlur).toHaveBeenCalledWith(undefined);
-        });
-
-        it("should trigger onBlur with value when tabbed out with value", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleBlur = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                value={new Date(2019, 0, 16, 12, 1, 12, 34)}
-                onBlur={handleBlur}
-              />
-              <label htmlFor="another-id">Another Field</label>
-              <input id="another-id" type="text" />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          expect(document.body).toHaveFocus();
-
-          // Tab in
-          userEvent.tab();
-          expect(element).toHaveFocus();
-
-          // Tab out
-          userEvent.tab();
-
-          // Should have triggered "onBlur"
-          expect(handleBlur).toHaveBeenCalledTimes(1);
-          expect(element).not.toHaveFocus();
-          expect(handleBlur).toHaveBeenCalledWith(
-            new Date(2019, 0, 16, 12, 1, 12, 34)
-          );
-        });
-
-        it("should trigger onBlur when picking a first date", async () => {
-          mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-          const handleBlur = vi.fn();
-
-          // Arrange
-          render(
-            <>
-              <label htmlFor="some-id">Some Field</label>
-              <DateTime
-                id="some-id"
-                dateFormat={FULL_DATE_FORMAT}
-                timeFormat={false}
-                onBlur={handleBlur}
-              />
-            </>
-          );
-
-          const element = await screen.findByLabelText("Some Field");
-          expect(element).toHaveValue("");
-          expect(
-            screen.queryByTestId("picker-wrapper")
-          ).not.toBeInTheDocument();
-
-          // Act
-          // Open picker
-          userEvent.click(element);
-
-          // Assert
-          expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-          const someDay = await screen.findByText("16");
-          expect(someDay).toBeVisible();
-
-          // Pick date
-          userEvent.click(someDay);
-
-          expect(element).toHaveValue("01/16/2019");
-
-          expect(handleBlur).toHaveBeenCalledTimes(1);
-          expect(handleBlur).toHaveBeenCalledWith(
-            new Date(2019, 0, 16, 0, 0, 0, 0)
-          );
-        });
-      });
-    });
-
-    describe("keyboard", () => {
-      it("should let you type to mark a full date active", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FULL_DATE_FORMAT}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should start visible with nothing active
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-        expect(await screen.findByText("16")).toBeVisible();
-        expect(await screen.findByText("16")).not.toHaveClass("rdtActive");
-
-        userEvent.type(element, "06/16/2015");
-
-        // Assert the typed value is now active
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-        expect(await screen.findByText("16")).toBeVisible();
-        expect(await screen.findByText("16")).toHaveClass("rdtActive");
-      });
-
-      it("should let you type to mark a month/year active", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={`${FORMATS.MONTH}/${FORMATS.YEAR}`}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should start visible with nothing active
-        expect(await screen.findByTestId("month-picker")).toBeVisible();
-        expect(await screen.findByText(/jun/i)).toBeVisible();
-        expect(await screen.findByText(/jun/i)).not.toHaveClass("rdtActive");
-
-        userEvent.type(element, "06/2015");
-
-        // Assert the typed value is now active
-        expect(await screen.findByTestId("month-picker")).toBeVisible();
-        expect(await screen.findByText(/jun/i)).toBeVisible();
-        expect(await screen.findByText(/jun/i)).toHaveClass("rdtActive");
-      });
-
-      it("should let you type to mark a year active", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FORMATS.YEAR}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should start visible with nothing active
-        expect(await screen.findByTestId("year-picker")).toBeVisible();
-        expect(await screen.findByText(/2015/i)).toBeVisible();
-        expect(await screen.findByText(/2015/i)).not.toHaveClass("rdtActive");
-
-        userEvent.type(element, "2015");
-
-        // Assert the typed value is now active
-        expect(await screen.findByTestId("year-picker")).toBeVisible();
-        expect(await screen.findByText(/2015/i)).toBeVisible();
-        expect(await screen.findByText(/2015/i)).toHaveClass("rdtActive");
-      });
-
-      it("should let you type to mark a time active", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={false}
-              timeFormat={FULL_TIME_FORMAT}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should start visible with nothing active
-        const picker = await screen.findByTestId("time-picker");
-        expect(picker).toBeVisible();
-        {
-          const textContent = picker.textContent?.replace(/\W+/g, "");
-          expect(textContent).toMatch(/1200AM/i);
-        }
-
-        userEvent.paste(element, "4:13 PM");
-
-        // Assert the typed value is now active
-        expect(await screen.findByTestId("time-picker")).toBeVisible();
-        {
-          const textContent = picker.textContent?.replace(/\W+/g, "");
-          expect(textContent).toMatch(/413PM/i);
-        }
-      });
-
-      it("should let you type to mark a date/time active", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FULL_DATE_FORMAT}
-              timeFormat={FULL_TIME_FORMAT}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should start visible with nothing active
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-        expect(await screen.findByText("16")).toBeVisible();
-        expect(await screen.findByText("16")).not.toHaveClass("rdtActive");
-
-        userEvent.type(element, "06/16/2015 12:00 AM");
-
-        // Assert the typed value is now active
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-        expect(await screen.findByText("16")).toBeVisible();
-        expect(await screen.findByText("16")).toHaveClass("rdtActive");
-      });
-
-      it("should show when tabbed in", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FULL_DATE_FORMAT}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        expect(document.body).toHaveFocus();
-
-        // Open picker
-        userEvent.tab();
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-        expect(element).toHaveFocus();
-
-        // Should become visible
-        expect(await screen.findByTestId("picker-wrapper")).toBeVisible();
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-      });
-
-      it("should hide when open and hitting enter", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FULL_DATE_FORMAT}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should become visible
-        expect(await screen.findByTestId("picker-wrapper")).toBeVisible();
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-        // Hit enter
-        userEvent.type(element, "{enter}");
-
-        // Assert the picker is closed
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-        expect(screen.queryByTestId("day-picker")).not.toBeInTheDocument();
-      });
-
-      it("should open when closed and hitting down arrow", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FULL_DATE_FORMAT}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Hit down arrow
-        userEvent.type(element, "{arrowdown}");
-
-        // Assert the picker is open
-        expect(await screen.findByTestId("picker-wrapper")).toBeVisible();
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-      });
-
-      it("should hide when open and hitting escape", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FULL_DATE_FORMAT}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should become visible
-        expect(await screen.findByTestId("picker-wrapper")).toBeVisible();
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-        // Hit escape
-        userEvent.type(element, "{esc}");
-
-        // Assert the picker is closed
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-        expect(screen.queryByTestId("day-picker")).not.toBeInTheDocument();
-      });
-
-      it("should hide when open and hitting tab", async () => {
-        mockDate(new Date(2019, 0, 1, 12, 1, 12, 34));
-
-        // Arrange
-        render(
-          <>
-            <label htmlFor="some-id">Some Field</label>
-            <DateTime
-              id="some-id"
-              dateFormat={FULL_DATE_FORMAT}
-              timeFormat={false}
-            />
-          </>
-        );
-
-        const element = await screen.findByLabelText("Some Field");
-        expect(element).toHaveValue("");
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-
-        // Act
-        // Open picker
-        userEvent.click(element);
-        expect(element).toHaveFocus();
-
-        // Should become visible
-        expect(await screen.findByTestId("picker-wrapper")).toBeVisible();
-        expect(await screen.findByTestId("day-picker")).toBeVisible();
-
-        // Hit tab
-        fireEvent.keyDown(element, {
-          key: "Tab",
-          code: 9,
-          keyCode: 9,
-          charCode: 9,
-        });
-        fireEvent.keyUp(element, {
-          key: "Tab",
-          code: 9,
-          keyCode: 9,
-          charCode: 9,
-        });
-
-        // Assert the picker is closed
-        expect(screen.queryByTestId("picker-wrapper")).not.toBeInTheDocument();
-        expect(screen.queryByTestId("day-picker")).not.toBeInTheDocument();
-      });
-    });
-  });
 });
